feat: add RuntimeError variant and enhance CLI functionality
